refactor: error page 위치 수정
