+++ Frontend/src/components/complete/CompleteMain/CompleteMain.style.tsx
@@ -26,7 +26,8 @@ export const GuideText = styled.div`
 export const CarImg = styled.div<{ $isExternal: boolean }>`
   width: 589px;
   height: 326px;
-  ${({ $isExternal }) => !$isExternal && `margin: 20px;`}
+  margin: 20px;
+  ${({ $isExternal }) => !$isExternal && ``}
 
   & > img {
     width: 100%;
