+++ Frontend/src/routes/Router.tsx
@@ -3,8 +3,7 @@ import MainPage from '../pages/MainPage/MainPage.tsx';
 import SelfModePage from '../pages/SelfModePage/SelfModePage.tsx';
 import GuideModePage from '../pages/GuideModepage/GuideModePage.tsx';
 import CompletePage from '@/pages/CompletePage/CompletePage.tsx';
-import PrivateRoute from './PrivateRoute.tsx';
-import ErrorPage from '@/pages/ErrorPage/ErrorPage.tsx';
+import ErrorPage from '@/components/complete/ErrorPage/ErrorPage.tsx';
 
 export default function Router() {
   return (
@@ -14,9 +13,7 @@ export default function Router() {
           <Route path="/" element={<MainPage />} />
           <Route path="/self-mode" element={<SelfModePage />} />
           <Route path="/guide-mode" element={<GuideModePage />} />
-          <Route element={<PrivateRoute />}>
-            <Route path="/complete" element={<CompletePage />} />
-          </Route>
+          <Route path="/complete" element={<CompletePage />} />
           <Route path="/error" element={<ErrorPage />} />
         </Routes>
       </BrowserRouter>
